transferred gene converter and trimming function back to import_data.py to avoid recursive import. Need to figure something else out
